------------------------------------------------------------------------ r3466 | drivi01 | 2010-08-16 12:45:12 -0400 (Mon, 16 Aug 2010) | 15 lines
Ingres installer SEGV on x64 while performing installation
at Services dialog.  The problem appears while executing
ingres_verify_user function which calls CheckUserRights
and is where the problem occurs.

It appears to be related to the NetApiBufferFree function
where pGroups variable is not pointing at the head of
the allocated enumeration list.

This change stores the head of the list, then allows
the function to enumerate through the list of groups
and then assigns the stored address back to pGroups
so it points at the beginning of the list again.

------------------------------------------------------------------------
r3467 | drivi01 | 2010-08-16 14:15:14 -0400 (Mon, 16 Aug 2010) | 12 lines

Ingres 64-bit installer is a new product.
There's no need to keep features around to prevent
major upgrade because this version of installer
is not going to be used for upgrades from previous
versions.

This change removes all ICE 64-bit ism files and
any trace of Ice feature existance from the
installer. It also removes it from the build
script for 64-bit installer only.

------------------------------------------------------------------------