------------------------------------------------------------------------ r3480 | schka01 | 2010-08-19 13:15:19 -0400 (Thu, 19 Aug 2010) | 15 lines
In a UTF8 installation, ordinary character-type columns
contain UTF8 data.  However, the normal character
type coercions don't know anything about UTF8, and
they do no checking, nor normalization.  This can lead
to invalid data being loaded, causing various query
problems later.  The observed problem was E_AD5010
during a CREATE INDEX because of non-UTF8 data.

This change causes COPY FROM to validate incoming
data as being valid UTF8 in a UTF8 installation,
when the target columns are ordinary character-type
columns.  It also normalizes before sending to
the back-end.

------------------------------------------------------------------------
r3481 | lunbr01 | 2010-08-19 16:30:12 -0400 (Thu, 19 Aug 2010) | 123 lines

On Windows, ingstart.exe may crash or delay for up to 30 seconds
after starting the name server if II_GC_PROT is set to "tcp_ip".
The problem is timing-related and tends to occur frequently on
some machines (VMs and single-processor) and never on others.

Small change, long explanation...

The delay occurs when ingstart is pinging the name server (iigcn)
to see if it is up before starting the other servers.  Ingstart
will do this up to 30 times with a sleep of 1 second after each
unsuccessful attempt.  The "ping" is essentially a connection
request (GCA_REQUEST) to the name server.  Any status returned
other than 0 (OK) or E_GC0032_NO_PEER is considered a failure
and the ping loop continues.  The problem was that, although the
name server had successfully started up, the status being returned
(on the problematic machines) was a 0x0001FE07 (NP_IO_FAIL).
When the loop limit of 30 was reached, if a crash hadn't yet
occurred, ingstart would continue and start up the rest of the
Ingres servers.

The bad status (0x0001FE07) was being returned when all local
communications I/O under the GCA_REQUEST completed immediately.
The flow, somewhat complicated, is as follows.
After the network connect completes successfully, a peerinfo
message is sent from the GCA CL layer and then another
(different) one from the GCA layer, followed by a receive
to get a peerinfo message back from the server (in this case
the name server).  The callback to the completion routine of each
action triggers the next action.  When tcp_ip is used as the local
protocol, if the network I/O completes immediately, then the tcp_ip
driver will call back the (gca cl) completion routine directly
before returning; that is, it doesn't return and then wait for the
I/O completion to make the callback, which is what happens if the
I/O initially is pending.  Named pipes, on the other hand, ALWAYS
returns (to GCsync() or similar function) after issuing the network
I/O and then lets the I/O completion do the callback.  The immediate
completion logic can cause the calls to the driver to get nested
since calling the completion routine will often call back into the
driver for the next action. This indeed usually occurs for the 2
peerinfo sends and sometimes for the peerinfo receive.  More
specifically, gcacl.c function GCDriverRequest_exit(), which is
only called if II_GC_PROT=tcp_ip, calls GCrequestSendPeer() to start
the peerinfo exchange.  If a non-zero status is returned, then it
assumes that a major error occurred and it then calls the completion
routine itself; if no error, it assumes the lower layers will call
(or have called) the completion routine (with or without an error),
so it just returns.  The problem was that GCrequestSendPeer() was
erroneously returning status 0x000C0032 (E_GC0032_NO_PEER) when all
3 peerinfo actions completed immediately.  They all shared the same
svc_parms and the receive of peerinfo from a name server sets
svc_parms->status to E_GC0032_NO_PEER (as it is expected to do and
is not an error).  GCrequestSendPeer() was erroneously returning
svc_parms->status as his status after calling GCDriverSendPeer().
If all I/Os completed immediately, svc_parms->status was
E_GC0032_NO_PEER, which was then picked up by caller
GCDriverRequest_exit() as an error, which it shouldn't have been.
3 bad things would then happen as as result.
 -  First, the iocb (GCA CL IO control block for the connect)
    would get MEfree'd but the code dropped down and set
    asipc->connect_iocb=iocb; the memory would then get
    freed again shortly after that in GCrelease().  Hence,
    the 2nd MEfree() would often free memory that had been
    reallocated for another purpose, which is what sometimes
    led to a crash.
 -  The 2nd bad thing that happened was the completion routine
    would be called when the lower layer had already called it.
    This caused the GCA state machine to get confused and issue
    a GCA_SEND of 0 bytes...not specifically causing a problem,
    but definitely undesired.
 -  The last (3rd) bad thing was that GCDriverRequest_exit()
    set svc_parms->status to NP_IO_FAIL due to the non-zero status
    of E_GC0032_NO_PEER and this percolated back up to the ping
    iigcn code in ingstart, so the code thought the name server
    was still not up.  This caused a delay in startup, because
    ingstart would then try again (and again) up to 30 times
    (with 1 second sleep each).

The fix consisted of the following:
1. In GCrequestSendPeer(), just return OK, not svc_parms->status,
   after calling GCDriverSendPeer().  The latter doesn't return a
   status itself, which would have been the thing to return if it did.
   At this point, the lower layers will call back the completion
   routine (with or without an error) so there is no point in
   setting a non-zero status and causing the caller to also call
   the completion routine.
2. Remove the MEfree's of the iocb after errors are detected in
   GCDriverRequest_exit() and move the saving of the pointer to it
   to earlier in the code; namely, when it is initially allocated
   in GCrequest().  The memory will then get freed in only one place
   (in GCrelease()) and only one time, regardless if there are errors
   or not.  This makes the code a bit simpler and less error-prone;
   it had previously been modeled after similar code in the pipes
   logic, which should also probably be changed similarly someday,
   ...though, if it ain't broke...
3. Added a debug trace statement if an error status is ever
   returned in this area of code (after call to GCrequestSendPeer()).

Another item fixed, though it didn't cause a problem in this
particular situation, was that a flag in the driver that indicated
it was shutting down was not getting reset (turned back off)
whenever the driver went through initialization.  Unlike most
applications and servers which initialize and terminate
the driver layer only once, ingstart (due to the ping_iigcn())
logic, would serially do it (init and term) multiple times
(until the name server was thought to be up or loop limit of
30 reached).  The ping_iigcn() code in the ingstart loop does
a full GCA_INITIATE, GCA_REQUEST, GCA_TERMINATE sequence.
Thus, when it doesn't think the name server is up right away,
the code will cause the driver to initialize and shutdown
multiple times.  The shutting down flag (static variable) was
initialized as a static variable to FALSE and then set to TRUE in
termination (GCwinsock2_exit()).  On subsequent initializations,
it was not getting reset back to FALSE.  The change now sets it to
FALSE in initialization (GCwinsock2_init()).  The flag is only
relevant (currently) if an I/O completion comes in without an
overlapped structure, which should only occur in a real shutdown.
Hence, it wasn't causing a problem, per se, but I wanted to
clean it up because it conceivably could.
Arguably, ingstart probably doesn't need to init and term GCA
in the ping loop, but it should be allowed to do so; hence,
I did not modify the ingstart code.

------------------------------------------------------------------------