------------------------------------------------------------------------ r3499 | drivi01 | 2010-08-23 13:15:15 -0400 (Mon, 23 Aug 2010) | 25 lines
In VDBA DOM, when connected to Ingres VectorWise
installations, in Optimizedb dialog under Specify
Tables, the list of tables displayed filters out
VectorWise tables.

This is due to the fact that the same dialog, which
is actually an AuditDB dialog is used by many other
commands such as usermod, optimizedb, "create alarm",
and etc. to display the list of tables, however,
most of those commands aren't supported for Ingres
VectorWise and we filter out Ingres VectorWise tables
from the list of displayed tables.  However, optimizedb
is supported on VectorWise tables but uses the same
dialog and inadvertently gets a filtered list of tables.

This change adds some additional functions to AuditDB
code to allow optimizedb to take a slightly different
execution path.  It also adds a 3rd parameter to the
function that loads the tables into the list to
specify what type of tables we'd like to see in the list.
This fixes the list of tables for optimizedb to include
all tables, but preserves filtered list for other
commands/dialogs.

------------------------------------------------------------------------
r3500 | stial01 | 2010-08-23 18:45:15 -0400 (Mon, 23 Aug 2010) | 8 lines

SIR 121123 Long IDs
Servicedesck 146328 E_EQ0216 Maximum word length, 64, is exceeded by
'long-id....'

Description of changes:
Increase constants in eqscan.h to hold long ids.

------------------------------------------------------------------------