------------------------------------------------------------------------ r3503 | lunbr01 | 2010-08-24 11:00:11 -0400 (Tue, 24 Aug 2010) | 53 lines
On Windows, IVM crashes in find_procs() if II_GC_PROT=tcp_ip.
This is because the IVM code built on Windows depends on pipe names,
rather than TCP/IP ports, in iinamu "show servers" output.  That is,
the code depended on the installation id and server type being
embedded in the listen address and separated by backslash (\);
for example, II\NMSVR\14a8.  Code should be generic and not dependent
on format of local listen ports, which "ifdef MAINWIN" code already is.
NOTE that "#ifdef MAINWIN" had been used to generate the Unix version
of the code.  Hence, the fix is to keep the MAINWIN code and strip
out the non-MAINWIN code that was used to detect and parse pipe
names.  MAINWIN-specific code that was truly for Unix vs Windows
differences and had nothing to do with named pipe listen address
format was left "as is".

As a result of removing non-MAINWIN code, some code simplification
was possible.  There were also a couple of problems in the Unix/MAINWIN
version of the code which were fixed, such as code that searched
for the instance string (the local listen address) in iinamu "show
servers" output.  Some simplifications done were:
1. Removed FormatKey2Find() and replaced calls to it with simple
   sprintf().
2. Removed code checking for master name server (Ingres undoc'd
   "discovery" feature) since remote name servers do not show
   up in iinamu "show servers" output and, hence, do not need
   to be skipped.

NOTE: One piece of code in getinst.cpp that is Windows
"named pipe"-specific was left in so as not to lose "minor"
existing functionality that looks for "lost" instances (rare).
The code will not work in the general case (like tcp/ip for
local IPC), which is why it was #ifndef'd.  To fix for general
case would likely require a major change to the flow of the
entire function code in find_procs(); for example, by processing
iinamu output once sequentially rather than repeatedly
"by server/component" and blanking out used instances in iinamu
output to keep track of what had been processed.
If II_GC_PROT=tcp_ip on Windows, the code is essentially
a no-op and behaves like Unix/tcp_ip local IPC where the
code is not even generated.

Suggestion for future (which was beyond scope of this fix):
In getinst.cpp function find_procs(), rewrite flow to process
iinamu output sequentially one time, processing each line for
the appropriate server appropriately.  This would eliminate
the need to blank out listen addresses in the iinamu output
when they had been processed, eliminate scans of iinamu output
for each server type, and enable the "lost instance" code to
work in all cases.  Additionally, put all the server keys and
info in a server table to avoid all the redundant code in this
section; this would also make it easier to add and delete server
types in the future.

------------------------------------------------------------------------
r3504 | thogo01 | 2010-08-24 18:00:12 -0400 (Tue, 24 Aug 2010) | 14 lines

Added registration bit for DAS MIB so that IMA will not
ignore GCD servers.

Skip NMSVR mib registrations.  These are Name Servers
from other installations or a duplicate of the local
IINMSVR Name Server registration.

Fixed display of GCB registration bit and added NMSVR
and GCD bit display.

Minor cleanup in GCD to ensure API shutdown occurs fully.

------------------------------------------------------------------------