------------------------------------------------------------------------ r3507 | bonro01 | 2010-08-25 16:30:15 -0400 (Wed, 25 Aug 2010) | 3 lines
Remove the confusing dual logging device prompt from the installation scripts.

------------------------------------------------------------------------
r3508 | jenjo02 | 2010-08-25 17:30:11 -0400 (Wed, 25 Aug 2010) | 10 lines

The LGC area of the DMP_LCTX allocated for use by
page undo is uninitialized. If that memory looks
sufficiently like a LGC to LG_position, it may
cause its initialization to be bypassed and may
lead to a SEGV in LG_position.

This change fills the LGC area with zeroes prior
to positioning the log.

------------------------------------------------------------------------