------------------------------------------------------------------------ r3508 | jenjo02 | 2010-08-25 17:30:11 -0400 (Wed, 25 Aug 2010) | 10 lines
The LGC area of the DMP_LCTX allocated for use by
page undo is uninitialized. If that memory looks
sufficiently like a LGC to LG_position, it may
cause its initialization to be bypassed and may
lead to a SEGV in LG_position.

This change fills the LGC area with zeroes prior
to positioning the log.

------------------------------------------------------------------------
r3509 | rajus01 | 2010-08-25 17:30:19 -0400 (Wed, 25 Aug 2010) | 7 lines

The DBMS server takes program exception during 'ingstart' due to misconfiguration of GCF/GCN security parameters. SIR 120874 enhanced debugging in DBMS. As a result the CL_ERR_DESC structure got changed to include errfile and errline. During 'ingstart' the Name Server fails to start due to misconfigured security mechanism. The GCN startup failure is ignored during 'ingstart' and it continues to start the DBMS server. The DBMS server checks for errline and errfile which were pointing to some bogus memory area thus resulting in SEGV when formatting the error message for logging.

Upon start up of the recovery server it issues a GCA_INITIATE request and this request fails with GC1003 error in the DBMS during GCS initialization and a CL/OS request was not made yet. Further GCA/DBMS research indicated that GCA parameter 'gca_os_status' of type CL_ERR_DESC needs to be zero filled prior to use.

The change is to use CL_CLEAR_ERR macro to clear the 'gca_os_status' value during GCA service request. Clearing this value eliminated the annoying ULE_FORMAT error and program exception in the DBMS.

------------------------------------------------------------------------