------------------------------------------------------------------------ r3510 | miket | 2010-08-26 07:32:52 -0400 (Thu, 26 Aug 2010) | 13 lines
Trac Number: #584

Bug Number: 122403

Description of changes:
Pre-GA build fixes
(1) SI fix insert, update, delete (build was the only operation fully working)
(2) SI hash on encrypted value for security
(3) disable (currently broken) MODIFY INDEX HASH to HASH
(4) improve message saying that ALTER TABLE cannot be used with encryption
(5) improve error messages in dm1e.c
(6) remove some debugging and tracing code

------------------------------------------------------------------------
r3511 | horda03 | 2010-08-26 09:45:11 -0400 (Thu, 26 Aug 2010) | 54 lines

Review Board IP  http://reviewboard.ingres.prv/r/5595

Cross integration of change 507016

Since Ingres 9.3 the view IISEQUENCES has two new columns
unordered_flag and ident_flag. the COPYDB/UNLOADDB utilities
have been modified to make use of these flags when dealing
with sequences.

This causes a problem when the utility is used (via Ingres/Net)
against an earlier release of Ingres where these columns are
not defined. This IP modifies the query used based on the version
of the Ingres the utility is connected to, this is done by
replacing the EMBEDDED SQL with constructed query which is executed
via the EXECUTE IMMEDIATE function.

When connected to an earlier release of Ingres than 9.3, the
"unordered_flag" is treated as the constant 'N' which is the
equivalent sequence existing prior to an upgrade. Ditto for the
"ident_flag".

xfsequences()

Change the EXEC SQL DECLARE GLOBAL TEMPORARY TABLE session.seq_list
to an STprintf( tmp_tbl_cmd,"DECLARE...."). This allows the "unordered_flag"
column to be created for IISEQUENCES in a 9.3+ installation, set to
'N' (with the same column name) for earlier releases.

A similar method is used for inserting the rows into the temporary table,
but now all C variables used in the query are now inserted as strings
(using '%s'). Where iisequences doesn''t have an unordered_flag column,
the string constant 'N' is used.

There is another function with this file that makes use of the 'unordered_flag',
but this is only used for Identity sequences, which were introduced in Ingres 9.3,
so I don't anticipate the function ever being invoked when connected to an
earlier release.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------