------------------------------------------------------------------------ r3521 | stephenb | 2010-08-30 01:09:03 -0400 (Mon, 30 Aug 2010) | 20 lines
Bug Number: 124302

rns.start
     124302
     (GENERIC)
      selecting character result row fields from a db proc in a UTF8 install
      may result in a blank field
rns.end

Description of changes:
desc.start
The problem is caused by the fact that there is code in opc_resdoms() that tries
to fiddle with the output length of character fields for Unicode before calling
opc_cqual1. This function expects the length to be un-initialized when opr_offset is
set to OPR_NO_OFFSET. Fiddling with the length is un-necessary, it is fixed in PSF,
just remove the code.

Related Service Desk Issues: 146111
Build Considerations: NONE
desc.end
------------------------------------------------------------------------