------------------------------------------------------------------------ r3529 | horda03 | 2010-08-31 09:45:17 -0400 (Tue, 31 Aug 2010) | 29 lines
Review Board IP  http://reviewboard.ingres.prv/r/5655

Cross integration of change 507089

In my original change for Bug 124304 I converted an EXEC SQL
statment to an STprintf and an EXECUTE IMMEDIATE. But I forgot
to convert "(column_default_val like 'next value for%') AND"
because using STprintf, the '%' represents the start of a
format sequence, to get the '%' (SQL wildcard), need to add an
'%'.

This IP fixes this oversight.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------
r3530 | hanje04 | 2010-08-31 12:00:25 -0400 (Tue, 31 Aug 2010) | 19 lines

The values for some of the settings defined by the "Configuration Type"
dialog in the GUI installer differ from the values defined in the DDS.

    Transactional

    system_maxlocks = 1200, DDS=1500

    Business Intelligence

    dmf_group_size  = 8, DDS=32
    p8k.dmf_group_size  = 8, DDS=32

    Content Management

    rcp.lock.lock_limit = 1285000, DDS=325000

Protect lock_limit too as it's derived

------------------------------------------------------------------------