------------------------------------------------------------------------ r3530 | hanje04 | 2010-08-31 12:00:25 -0400 (Tue, 31 Aug 2010) | 19 lines
The values for some of the settings defined by the "Configuration Type"
dialog in the GUI installer differ from the values defined in the DDS.

    Transactional

    system_maxlocks = 1200, DDS=1500

    Business Intelligence

    dmf_group_size  = 8, DDS=32
    p8k.dmf_group_size  = 8, DDS=32

    Content Management

    rcp.lock.lock_limit = 1285000, DDS=325000

Protect lock_limit too as it's derived

------------------------------------------------------------------------
r3531 | jenjo02 | 2010-08-31 12:30:26 -0400 (Tue, 31 Aug 2010) | 13 lines

Changes to DMU for SIR 120874 inadvertantly
dropped the setting of dberr->err_data to the index
value of the location name for certain RAW
errors. QEF uses dberr->err_data to report the
name of the location of the failing RAW operation,
so it always reports the name of the first location
instead of the n-th.

Changed the function prototype of dm2u_raw_location_free()
to pass the index value for setting dberr->err_data
by the SETDBERR macro.

------------------------------------------------------------------------