------------------------------------------------------------------------ r3531 | jenjo02 | 2010-08-31 12:30:26 -0400 (Tue, 31 Aug 2010) | 13 lines
Changes to DMU for SIR 120874 inadvertantly
dropped the setting of dberr->err_data to the index
value of the location name for certain RAW
errors. QEF uses dberr->err_data to report the
name of the location of the failing RAW operation,
so it always reports the name of the first location
instead of the n-th.

Changed the function prototype of dm2u_raw_location_free()
to pass the index value for setting dberr->err_data
by the SETDBERR macro.

------------------------------------------------------------------------
r3532 | lunbr01 | 2010-08-31 13:00:32 -0400 (Tue, 31 Aug 2010) | 34 lines

On Windows, the closesocket() function is called during session
disconnect and is expected to do an implicit graceful shutdown
of the session.  However, the shutdown/disconnect is not always
graceful, and messages that have been sent but not yet received
by the other side are occasionally dropped.  This can result in
hung sessions and/or connection abort errorlog messages in Ingres
servers.

The problem can be difficult to reproduce, but occurs more
reliably when II_GC_PROT is set to tcp_ip and running through
Ingres/Net.  The message dropped, when it occurs, is generally
the GCA_RELEASE message from the server-side GCC to the DBMS.
Theoretically, the problem could affect any TCP/IP connection
in Ingres, not just the GCC to DBMS connection.

The fix is to precede the closesocket() call in GCC_DISCONNECT
in the tcp_ip protocol driver with a DisconnectEx() call.
This forces a graceful disconnect, rather than relying on
closesocket()'s implicit graceful disconnect, which doesn't
always work and can cause sent messages to occasionally be dropped
before being received by the other side of the connection.
Microsoft documentation states that closesocket() will do an
implicit graceful shutdown of a session by default.  However,
the documentation goes on to officially prescribe using SO_LINGER
socket option or the shutdown() function to ensure graceful
connection shutdowns.  A comment to the documentation then states
that SO_LINGER doesn't work and that DisconnectEx() is the only
way that always works, which is why it was chosen here.

NOTE that DisconnectEx() is a Winsock extended function and
so its address must be found by calling WSAIoctl with option
SIO_GET_EXTENSION_FUNCTION_POINTER.

------------------------------------------------------------------------