------------------------------------------------------------------------ r3532 | lunbr01 | 2010-08-31 13:00:32 -0400 (Tue, 31 Aug 2010) | 34 lines
On Windows, the closesocket() function is called during session
disconnect and is expected to do an implicit graceful shutdown
of the session.  However, the shutdown/disconnect is not always
graceful, and messages that have been sent but not yet received
by the other side are occasionally dropped.  This can result in
hung sessions and/or connection abort errorlog messages in Ingres
servers.

The problem can be difficult to reproduce, but occurs more
reliably when II_GC_PROT is set to tcp_ip and running through
Ingres/Net.  The message dropped, when it occurs, is generally
the GCA_RELEASE message from the server-side GCC to the DBMS.
Theoretically, the problem could affect any TCP/IP connection
in Ingres, not just the GCC to DBMS connection.

The fix is to precede the closesocket() call in GCC_DISCONNECT
in the tcp_ip protocol driver with a DisconnectEx() call.
This forces a graceful disconnect, rather than relying on
closesocket()'s implicit graceful disconnect, which doesn't
always work and can cause sent messages to occasionally be dropped
before being received by the other side of the connection.
Microsoft documentation states that closesocket() will do an
implicit graceful shutdown of a session by default.  However,
the documentation goes on to officially prescribe using SO_LINGER
socket option or the shutdown() function to ensure graceful
connection shutdowns.  A comment to the documentation then states
that SO_LINGER doesn't work and that DisconnectEx() is the only
way that always works, which is why it was chosen here.

NOTE that DisconnectEx() is a Winsock extended function and
so its address must be found by calling WSAIoctl with option
SIO_GET_EXTENSION_FUNCTION_POINTER.

------------------------------------------------------------------------
r3533 | drivi01 | 2010-08-31 15:15:32 -0400 (Tue, 31 Aug 2010) | 14 lines

VDBA crashes when application exits if DOM window is opened.

This is due to the fact that in the place of an old ICE entry
in the tree an empty bogus record is being added.

This change removes comments around ice and puts preprocessor
defines around the record and the code responsible
for addition of the bogus record to the tree.

Preprocessor flag is a better way of diabling the ICE code
in VDBA because the code can be enabled at any time by
compiling with the BUILD_ICE preprocessor flag.

------------------------------------------------------------------------