------------------------------------------------------------------------ r3533 | drivi01 | 2010-08-31 15:15:32 -0400 (Tue, 31 Aug 2010) | 14 lines
VDBA crashes when application exits if DOM window is opened.

This is due to the fact that in the place of an old ICE entry
in the tree an empty bogus record is being added.

This change removes comments around ice and puts preprocessor
defines around the record and the code responsible
for addition of the bogus record to the tree.

Preprocessor flag is a better way of diabling the ICE code
in VDBA because the code can be enabled at any time by
compiling with the BUILD_ICE preprocessor flag.

------------------------------------------------------------------------