------------------------------------------------------------------------ r3551 | schka01 | 2010-09-03 07:15:13 -0400 (Fri, 03 Sep 2010) | 25 lines
The original REPEATED query plan caching mechanism enforces
a single call site for each query, and length variations
cannot occur except for variable-length strings.  There is
a mechanism to deal with variable-length strings in the
query plan.  This mechanism does not extend to any fixed-
length types, nor did it need to.

With the introduction of cache_dynamic, actual parameters can
be supplied by many different call sites, and so parameter
length variation is a possibility.  It all continues to work
for variable-length strings, but for fixed length parameters
it's necessary to verify that the lengths compiled into the
the cached QP match the lengths of the actual params.
(For DECIMAL we need to compare precision/scale as well.)

The cache_dynamic code was checking actual parameter data
types against the cached QP, but not the lengths.  Add
the missing check.

To make debugging a little easier, add trace point PS152
to turn on some cache_dynamic debug output that was
formerly under xDEBUG compile conditional.  The debug
output goes to the DBMS log.

------------------------------------------------------------------------
r3552 | kiria01 | 2010-09-03 09:15:42 -0400 (Fri, 03 Sep 2010) | 6 lines

Correct typo in member assignment

Correct the assignment.

------------------------------------------------------------------------