------------------------------------------------------------------------ r3561 | maspa05 | 2010-09-07 11:15:24 -0400 (Tue, 07 Sep 2010) | 40 lines
Review Board IP  http://reviewboard.ingres.prv/r/5682

SIR 124345 - add QEP segmented output (see bug 124274) to SC930
SIR 124346 - add EQY, end query record to SC930 with rowcount and error code

Changed the ult functions that check and set the SC930 status to use an i4
bitmask rather than a series of bools. As it was we had 2 variables and 2
sets of two functions - one for tracing and one for QEP. This way we can add
extra features to the SC930 without adding these functions each time.

Consequently ult_trace_qep() and ult_set_trace_qep() have been removed.
Relevant entries in iilibulf.def commented out for Windows build.

Amended existing calls to ult_always_trace() to check the relevant flag value,
usually SC930_TRACE for tracing ON.

Amended scf_trace to change the way SC930 is set. Added code so that set
notrace point sc930 turns off tracing. SC930 0 also turns tracing off. Can
also turn on QEPs and off, and choose the style. The default style is the
new segmented output. Also updated SC1000 output to give summary of SC930
options.

Ammended scs_sequencer to output the end query record. This includes the
rowcount and error code. Amended scs_qef_error to save the more informative
error code it has (otherwise we gets lots of E_QE0025 generic user errors).

Amended uleformat to save the error code specifically for SC930 tracing. This
was because it was proving difficult to track all the various translations of
different error codes in different facilities. This way we just get the last
error raised against this session.

Added cs_sc930_err to the CS_SCB to keep this error code in.

Added check in uldprtree to see if we're being called for an SC930 trace
and if so, and SC930_QEP_SEG is set then set the flag to output in
segmented style.

Incremented the SC930_VERSION to 8.

------------------------------------------------------------------------
r3562 | loera01 | 2010-09-07 11:15:37 -0400 (Tue, 07 Sep 2010) | 22 lines

Cross-integration of change 507187 from ingres10.

Change 507179 broke non-Windows builds due to extraneous "##" symbols
preceding the comment delimiters of the history flower box.  In addition,
unicode.c would not compile due to incompatibilty of the function template
with the include file.

I had added the "##" symbols since this worked on Linux: the comment
delimiters were not written to the sql.h and sqlext.h files.  On 64-bit
Solaris, the "##" indicators were retained, which broke ODBC builds.

In sqlext.h, I had added a new function template for SQLColAttributeW(), which
used SQLLEN arguments for the _WIN64 precompiler definition and SQLINTEGER
otherwise.  This matched the MS versio of sqlext.h.  I had not done the same
for the function declaration, which caused the function declaration to be at
odds with the template.   Thus, unicode.c would not compile.

I removed the unneeded "##" symbols and revised the declaration of
SQLColAttributeW() to match sqlext.h.

------------------------------------------------------------------------