------------------------------------------------------------------------ r3562 | loera01 | 2010-09-07 11:15:37 -0400 (Tue, 07 Sep 2010) | 22 lines
Cross-integration of change 507187 from ingres10.

Change 507179 broke non-Windows builds due to extraneous "##" symbols
preceding the comment delimiters of the history flower box.  In addition,
unicode.c would not compile due to incompatibilty of the function template
with the include file.

I had added the "##" symbols since this worked on Linux: the comment
delimiters were not written to the sql.h and sqlext.h files.  On 64-bit
Solaris, the "##" indicators were retained, which broke ODBC builds.

In sqlext.h, I had added a new function template for SQLColAttributeW(), which
used SQLLEN arguments for the _WIN64 precompiler definition and SQLINTEGER
otherwise.  This matched the MS versio of sqlext.h.  I had not done the same
for the function declaration, which caused the function declaration to be at
odds with the template.   Thus, unicode.c would not compile.

I removed the unneeded "##" symbols and revised the declaration of
SQLColAttributeW() to match sqlext.h.

------------------------------------------------------------------------
r3563 | hanal04 | 2010-09-07 11:15:52 -0400 (Tue, 07 Sep 2010) | 22 lines

Cross integration of change 506984

Review Board IP  http://reviewboard.ingres.prv/r/5333

In Ingres 2.6 and earlier releases the following SQL
was valid:

SELECT 32 / '$1.2'

In 9.x and above this fails because of coercion rule cahnges
that resolve a number of other problems. However, this presents
a problem for customers that are trying to upgrade.

This change adds money_config, a new DBMS configuration parameter.
When set to ON SCONST (string constants) are checked for explicit
money string content (a valid leading or trailing currency symbol)
and if found we establish a money constant instead of a string
constant. This allows the above SQL to work as it did in Ingres 2.6
(and earlier releases).

------------------------------------------------------------------------