------------------------------------------------------------------------ r3563 | hanal04 | 2010-09-07 11:15:52 -0400 (Tue, 07 Sep 2010) | 22 lines
Cross integration of change 506984

Review Board IP  http://reviewboard.ingres.prv/r/5333

In Ingres 2.6 and earlier releases the following SQL
was valid:

SELECT 32 / '$1.2'

In 9.x and above this fails because of coercion rule cahnges
that resolve a number of other problems. However, this presents
a problem for customers that are trying to upgrade.

This change adds money_config, a new DBMS configuration parameter.
When set to ON SCONST (string constants) are checked for explicit
money string content (a valid leading or trailing currency symbol)
and if found we establish a money constant instead of a string
constant. This allows the above SQL to work as it did in Ingres 2.6
(and earlier releases).

------------------------------------------------------------------------
r3564 | hanal04 | 2010-09-07 11:15:59 -0400 (Tue, 07 Sep 2010) | 15 lines

Cross integration of change 507188

Review Board IP  http://reviewboard.ingres.prv/r/5598

Money strings such as '25.10DM' are treated as character
strings and changes to the coercion rules mean that many
ABF aplications will now fail with coercion errors.
This change adds the environment variable II_MONEY_COMPAT
which will look for money strings during the compilation
of ABF applications. This money strings will be converted
to decimal values in order to avoid coercion errors during
execution.

------------------------------------------------------------------------