------------------------------------------------------------------------ r3574 | schka01 | 2010-09-08 08:45:10 -0400 (Wed, 08 Sep 2010) | 11 lines
When pst_dbpshow creates a DBPINFO data structure, it
filled in the DBP name and owner, but left the rest of
the structure uninitialized.  There was no code to
initialize the db_mask flags anywhere else in the parser,
so random flags could be set depending on what was in
memory.  The SD issue symptom was caused by the
"I am a row-returning DBP" flag being accidently set
when attempting to define a recursive table
procedure.  The wrong error message occurred.

------------------------------------------------------------------------
r3575 | miket | 2010-09-08 09:47:27 -0400 (Wed, 08 Sep 2010) | 9 lines

Trac Number: #584

Bug Number: 122403

Description of changes:
Define MODIFY_BUCKET_PNO_TID8_SZ (size of bucket number, partition number,
and tid8) that may be appended to record buffers for MODIFY. Use this
in allocating RCB buffers and in MODIFY processing.

------------------------------------------------------------------------