------------------------------------------------------------------------ r3580 | drivi01 | 2010-09-08 18:45:14 -0400 (Wed, 08 Sep 2010) | 20 lines
On Certain machines, when user opens IIA or IEA applications,
they may get a crash after they open File Dialog to import/export
files and then cancel the IIA/IEA wizard.

I suspect this problem surfaced before in different scenarios
but was never fixed.

The problem is due to the fact that function _tgetenv is used
to retrieve USERPROFILE variable and then the path is stored
in a TCHAR pointer.  That pointer is then used in strcat
to concatenate "Documents" directory to it.  I suspect user gets
a violation when application is exited because of appended
string on the pointer buffer from tgetenv.  That memory is
never allocated for concatenation and the only time this becomes
apparent is when the application exits and the memory is being freed.

The proper way of doing this is to copy a buffer we get from _tgetenv
to another string before concatenating it with another string.

------------------------------------------------------------------------
r3581 | miket | 2010-09-09 11:00:40 -0400 (Thu, 09 Sep 2010) | 8 lines

Trac Number: #584

Bug Number: 122403

Description of changes:
Fix cut-and-paste error; associate with E_DM00AD and 9413 message E_QE0193
and not E_QE0192.

------------------------------------------------------------------------